Code refatoring
